Separate the README and the overview page
For the README to be PyPi-compliant, it cannot contain Sphinx extension
elements, such as `:py:class:`. For the documentation's "Overview" page
to still be detailed and descriptive, I decided to separate the two, so
the README.rst file is now a little simpler.

Also, not sure why PyPi doesn't allow two links to have the same label
text. I had to change "See here" to "See this page" so it wouldn't
conflict with another "see here".